+++ packages/app-config-writer/src/variants-config/generateVariantsConfig.ts
@@ -9,7 +9,7 @@ import type { ToolsLogger } from '@sap-ux/logger';
  * Add variants configuration to an app or project.
  *
  * @param basePath - the base path where the package.json and ui5.yaml is
- * @param yamlPath - the path where the ui5.yaml is
+ * @param yamlPath - path to the ui5*.yaml file passed by cli
  * @param logger - logger
  * @param fs - the memfs editor instance
  * @returns Promise<Editor> - memfs editor instance with updated files
@@ -23,7 +23,7 @@ export async function generateVariantsConfig(
     if (!fs) {
         fs = create(createStorage());
     }
+    await updateMiddlewares(fs, basePath, yamlPath, logger);
     await addVariantsManagementScript(fs, basePath, yamlPath, logger);
-    await updateMiddlewares(fs, basePath, logger);
     return fs;
 }