+++ packages/app-config-writer/test/unit/variants-config/package-json.test.ts
@@ -14,6 +14,8 @@ describe('addVariantsManagementScript', () => {
 
     const basePath = join(__dirname, '../../fixtures/variants-config');
     const yamlPath = 'path/to/my/ui5.yaml';
+    const customYaml = 'path/toTmy/customui5.yaml';
+
     beforeEach(() => {
         jest.clearAllMocks();
         fs = createFS(createStorage());
@@ -45,8 +47,8 @@ describe('addVariantsManagementScript', () => {
     });
 
     test('do not update script in package.json when it already exists and is up-to-date', async () => {
-        const fioriToolsConfig = join(basePath, 'up-to-date');
-        await addVariantsManagementScript(fs, fioriToolsConfig, yamlPath, logger);
+        const fioriToolsConfig = join(basePath, 'up-to-date-config');
+        await addVariantsManagementScript(fs, fioriToolsConfig, 'myCustomUI5.yaml', logger);
         expect(infoLogMock).toHaveBeenCalledWith(`Script 'start-variants-management' is already up-to-date.`);
         expect(debugLogMock).not.toHaveBeenCalledWith(`Script 'start-variants-management' written to 'package.json'.`);
     });
@@ -56,5 +58,18 @@ describe('addVariantsManagementScript', () => {
         await expect(addVariantsManagementScript(fs, openSourceConfig, yamlPath, logger)).rejects.toThrowError(
             'No RTA editor specified in ui5.yaml.'
         );
+    });
+
+    test('no package.json file', async () => {
+        const basePath = join(__dirname, '../../fixtures/a-folder-that-does-not-exist');
+        await expect(addVariantsManagementScript(fs, basePath, yamlPath, logger)).rejects.toThrowError(
+            `Script 'start-variants-management' cannot be written to package.json. File 'package.json' not found at ${basePath}`
+        );
+    });
+
+    test('set --config flag if default ui5.yaml is not used', async () => {
+        const fioriToolsConfig = join(basePath, 'up-to-date-config');
+        await addVariantsManagementScript(fs, fioriToolsConfig, 'myCustomUI5.yaml', logger);
+        expect(fs.readJSON(join(fioriToolsConfig, 'package.json'))).toMatchSnapshot();
     });
 });