+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/app-config-writer
 
+## 0.4.50
+
+### Patch Changes
+
+-   fb26f92: Enhance RTA handling for variants-config command
+-   Updated dependencies [fb26f92]
+    -   @sap-ux/project-access@1.28.6
+    -   @sap-ux/axios-extension@1.17.0
+
+## 0.4.49
+
+### Patch Changes
+
+-   Updated dependencies [0120dda]
+    -   @sap-ux/axios-extension@1.17.0
+
 ## 0.4.48
 
 ### Patch Changes