+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/project-access
 
+## 1.28.6
+
+### Patch Changes
+
+-   fb26f92: Enhance RTA handling for variants-config command
+
 ## 1.28.5
 
 ### Patch Changes