+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.0.19
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.5
+-   @sap-ux/ui5-library-writer@0.5.33
+
+## 0.0.18
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-writer@0.5.32
+
 ## 0.0.17
 
 ### Patch Changes